Edited packing, renamed config file The configuration file has been renamed from "todo.json" to "conf.json" because "todo.json" is the name of all to-do lists created by ProjectTo-Do.  Also, the packing is now more convenient.
I need to find a way to edit the preferences inside of the program, but I have no idea how to make it work with the current interface!  Where could a button fit in?  I'll try to look into it, but I have no idea how I'm going to make it work.